fix(web): Improve cause of issue
One reason for the sometimes small text on the spacebar besides the
too small font size on `kmw-spacebar-caption` was that the calculation
of the ideal font size added double padding: it added two pixels of
padding to the height, but also multiplied the height value by 0.9 which
effectively added more padding. This change removes the 2 pixel padding
from the calculation.

The existing calculation varies the text size to fit in the available
width of the spacebar, so calculating with font size of 1em is reasonable
and better than the previous 0.6em. The previous value would lead to even
an smaller font size for long texts, far below the size that would be
sufficient to fit the text.