fix(developer): make sourcePath calculation for kmc-keyboard-info more lenient
Fixes #10542.

The `sourcePath` attribute should only be added to a .keyboard_info file
if the file is actually in an expected folder matching a pattern like
`(release|legacy|experimental)/k/keyboard`.

If the path does not match that pattern, then we just omit the field in
the emitted .keyboard_info or .model_info file.