fix(developer): keyboard that targets only web can cause crashes and other problems
Fixes #10655.
Fixes #10663.

If a .kmn file had only a .js target, then Keyman Developer would return
an empty string for the target (based on .kmx). This meant that the path
for the compiler was an empty string, the path for opening the build
folder was an empty string.

Refactored the relevant code to simplify the logic a little.