feat(windows): handle CRLF specific to windows
Added a function to replace `\r\n` with `\n` that is
U+000D U+000A with U+000A. And another to do the reverse
that is insert '\r` infront of any `\n` charcters to support
the CRLF for windows. Supporting unit tests also added.
The function have been integerated into the code and are used
before the set_if_need Keyman Core call and then in the processing
of actions returned from the Core.