fix(developer): use `--no-dedupe` flag to avoid HTTP 414 in sentry-cli
Ref https://github.com/getsentry/sentry-cli/issues/1657

Seeing requests with > 16K URI (e.g. 48KB!) like:

  GET https://sentry.io/api/0/projects/keyman/keyman-developer/releases/release@17.0.273-beta-test-10798/files/?cursor=&checksum=004512221481ac19e94117147b07f028ba6c31d4&...

This will skip the de-dupe phase which is intended to reduce the number
of files being uploaded; as it was failing in some cases, and thus
uploading the files anyway, I am hopeful that it will have no
significant impact on the build time.