refactor(developer): cleanup kmn compiler message namespaces
This is in preparation for message documentation.

Moves all messages with the KmnCompiler namespace into
KmnCompilerMessages. For now, keeps CompilerMessages class as a subclass
of KmnCompilerMessages, for removal in 18.0.

Moves messages with KmnCompiler namespace from kmw-compiler-messages.ts
to kmn-compiler-messages.ts. This avoids confusion as to source of
message, and should allow us in the future to split the classes
entirely.