fix(web): Fix layout for subkey menus with unusually-sized keys
The previous code didn't check that the keys we put in a subkey menu
actually fit within the width of the osk. Additionally the subkey menu
was limited to two rows. This change checks the total width of the keys
in a row. If it gets too wide we continue in the next row.

Fixes #10126.