chore(web): fixes pred-text karma config
The upgrade to chai means that karma's 'framework' feature is no longer compatible, meaning a _little_ extra legwork is required to use it for karma browser tests.  It's just a little bit, though.

Will need to replicate this for the other packages that use karma to test, as the pattern kinda proliferated.