change(common): preliminary update of min node version
This updates the minimum required node version to Node 20. The latest
version of Node 18 still contains the npm bug
(https://github.com/npm/cli/issues/7072) whereas Node 20 got updated to
a npm version that contains a fix. Node 20 is known to work with our
current code, so this change updates to that as an intermediate step
before we investigate if we can update to Node 22 as discussed for
Keyman 18.