chore(common): add test for interactions between deps and children
Fixes #11428.

It is possible for a dep to have child builds, and vice-versa. When
a dep has child builds, we want to avoid those child builds being run
twice in the scenario that the dep is also referred to directly (though
the presence of complex dependencies like this may be a red flag for our
dependency map, but that's a separate problem...)

This test does not currently fix the problem, but it starts the
exploration process.