chore(windows): remove cleanup-alpha-tasks
Remove the CleanupAlphaTasks that was specific to task
schedules created it for 15.0 alpha. Scheduled tasks naming
was changed to include a user name in the task. The clean up
function was to make sure the previous named tasks were removed.
Now there should be no 15.0 alpha users it can be removed.