fix(linux): move environment to separate job and checkout to separate dir
In #11571 we tried to fix the failing API verification by moving
setting up the environment and restoring artifacts to the same job
as the api verification itself. However, this doesn't help and the
`artifacts` subdirectory still vanishes, caused by
https://github.com/actions/checkout/issues/1201.

This change reverts the change introduced in #11571 and changes the
git checkout to end in a separate subdirectory.

Reverts: #11571