fix(windows): check IM window will be in a visible location
The IM display call back is used to show the IM window. With this change
deadcode is removed and a check is made to ensure the IM window is
completley visible coordinates for the monitor that has focus. If not
it will be position close to where the caret in the application that has
focus.